+++ libraries/core/src/main/kotlin/com/bumble/appyx/core/node/ParentNode.kt
@@ -169,13 +169,13 @@ abstract class ParentNode<NavTarget : Any>(
     }
 
     /**
-     * attachWorkflow executes provided action e.g. backstack.push(NodeANavTarget) and waits for the specific
+     * attachChild executes provided action e.g. backstack.push(NodeANavTarget) and waits for the specific
      * Node of type T to appear in the ParentNode's children list. It should happen almost immediately because it happens
      * on the main thread, but the order of actions is not preserved as lifecycleScope uses Dispatchers.Main.immediate.
      * As the result we're doing it asynchronously with timeout after which exception is thrown if
      * expected node has not appeared in the children list.
      */
-    protected suspend inline fun <reified T : Node> attachWorkflow(
+    protected suspend inline fun <reified T : Node> attachChild(
         timeout: Long = ATTACH_WORKFLOW_SYNC_TIMEOUT,
         crossinline action: () -> Unit
     ): T = withContext(lifecycleScope.coroutineContext) {
