+++ libraries/core/src/test/kotlin/com/bumble/appyx/core/node/NodeTest.kt
@@ -42,7 +42,7 @@ class NodeTest {
             private set
 
         suspend fun changeStatus(): TestNode {
-            return executeWorkflow {
+            return executeAction {
                 status = StatusExecuted
             }
         }
