+++ mkdocs.yml
@@ -35,6 +35,7 @@ nav:
           - Promoter: navmodel/promoter.md
           - Writing your own: navmodel/custom.md
       - Composable navigation: navigation/composable-navigation.md
+      - Explicit navigation: navigation/explicit-navigation.md
   - UI:
       - Children: ui/children-view.md
       - Transitions: ui/transitions.md
