+++ samples/sandbox/src/main/kotlin/com/bumble/appyx/sandbox/client/explicitnavigation/GrandchildNodeTwo.kt
@@ -1,5 +1,6 @@
-package com.bumble.appyx.sandbox.client.workflow
+package com.bumble.appyx.sandbox.client.explicitnavigation
 
+import android.util.Log
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
@@ -13,19 +14,22 @@ import androidx.compose.material.Button
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
-import androidx.compose.ui.Alignment.Companion.CenterHorizontally
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
 import com.bumble.appyx.core.modality.BuildContext
 import com.bumble.appyx.core.node.Node
-import com.bumble.appyx.sandbox.client.workflow.treenavigator.Navigator
+import com.bumble.appyx.sandbox.client.explicitnavigation.treenavigator.Navigator
 
-class GrandchildNodeOne(
+class GrandchildNodeTwo(
     buildContext: BuildContext,
     private val navigator: Navigator
 ) : Node(buildContext) {
+
+    suspend fun printLifecycleState() = executeAction<GrandchildNodeTwo> {
+        Log.e("Lifecycle", lifecycle.currentState.toString())
+    }
 
     @Composable
     override fun View(modifier: Modifier) {
@@ -41,7 +45,10 @@ class GrandchildNodeOne(
                     .wrapContentSize()
                     .align(Alignment.Center)
             ) {
-                Text(text = "Grandchild one", modifier = Modifier.align(CenterHorizontally))
+                Text(
+                    text = "Grandchild two",
+                    modifier = Modifier.align(Alignment.CenterHorizontally)
+                )
                 Spacer(modifier = Modifier.requiredHeight(8.dp))
                 Button(onClick = { navigator.navigateToChildOne() }) {
                     Text(text = "Navigate to child one")
