+++ samples/sandbox/src/main/kotlin/com/bumble/appyx/sandbox/client/explicitnavigation/treenavigator/Navigator.kt
@@ -1,0 +1,7 @@
+package com.bumble.appyx.sandbox.client.explicitnavigation.treenavigator
+
+
+interface Navigator {
+
+    fun navigateToChildOne()
+}
