--- samples/sandbox/src/main/kotlin/com/bumble/appyx/sandbox/client/workflow/treenavigator/Navigator.kt
@@ -1,7 +1,0 @@
-package com.bumble.appyx.sandbox.client.workflow.treenavigator
-
-
-interface Navigator {
-
-    fun navigateToChildOne()
-}
